improved script 028 to have 'run-new-only' mode

+++ scripts/_0_2_8_human_to_adm_comparison.py
@@ -29,14 +29,15 @@ AD_PROBES = {
     "DryRunEval-MJ5-eval": ['Probe 1', 'Probe 1-A.1', 'Probe 1-B.1', 'Probe 2', 'Response 2-A.1-B', 'Response 2-B.1-B', 'Response 2-B.1-B-gauze-u', 'Response 2-A.1-B-gauze-sp', 'Probe 2-A.1-A.1', 'Probe 2-B.1-A.1', 'Probe 2-A.1-B.1-A.1', 'Probe 2-B.1-B.1-A.1', 'Probe 3', 'Probe 4']
 }
 
-def main(mongoDB, EVAL_NUMBER=4):
+def main(mongoDB, EVAL_NUMBER=4, run_new_only=False):
     if EVAL_NUMBER == 5:
         ADEPT_URL = config('ADEPT_URL')
         ST_URL = config('ST_URL')
     text_scenario_collection = mongoDB['userScenarioResults']
     delegation_collection = mongoDB['surveyResults']
     comparison_collection = mongoDB['humanToADMComparison']
-    comparison_collection.delete_many({'evalNumber': EVAL_NUMBER})
+    if not run_new_only:
+        comparison_collection.delete_many({'evalNumber': EVAL_NUMBER})
     medic_collection = mongoDB['admMedics']
     adm_collection = mongoDB["test"]
     del_adm_runs_collection = mongoDB['delegationADMRuns']
@@ -52,7 +53,8 @@ def main(mongoDB, EVAL_NUMBER=4):
             continue
         session_id = entry.get('combinedSessionId', entry.get('serverSessionId'))
         pid = entry.get('participantID')
-        survey = list(delegation_collection.find({"results.Participant ID Page.questions.Participant ID.response": pid}))
+        # only get completed surveys!
+        survey = list(delegation_collection.find({"results.Participant ID Page.questions.Participant ID.response": pid, "results.Post-Scenario Measures": {'$exists': True}}))
         if len(survey) == 0:
             print(f"No survey found for {pid}")
             continue
@@ -68,6 +70,20 @@ def main(mongoDB, EVAL_NUMBER=4):
                     if adm is None:
                         continue
                     adm_session = adm['history'][len(adm['history'])-1]['parameters']['session_id']
+                    document = {
+                        'pid': pid,
+                        'adm_type': survey['results'][page]['admAlignment'],
+                        'text_scenario': scenario_id,
+                        'text_session_id': session_id,
+                        'adm_scenario': page_scenario,
+                        'adm_session_id': adm_session,
+                        'adm_author': survey['results'][page]['admAuthor'],
+                        'adm_alignment_target': survey['results'][page]['admTarget'],
+                        'evalNumber': EVAL_NUMBER
+                    }
+                    if run_new_only and does_doc_exist(comparison_collection, document):
+                        # do not rerun!
+                        continue
                     # create ST query param
                     query_param = f"session_1={session_id}&session_2={adm_session}"
                     for probe_id in ST_PROBES[scenario_id]:
@@ -78,18 +94,7 @@ def main(mongoDB, EVAL_NUMBER=4):
                     res = requests.get(f'{ST_URL}api/v1/alignment/session/subset?{query_param}').json()
                     # send document to mongo
                     if res is not None and 'score' in res:
-                        document = {
-                            'pid': pid,
-                            'adm_type': survey['results'][page]['admAlignment'],
-                            'score': res['score'],
-                            'text_scenario': scenario_id,
-                            'text_session_id': session_id,
-                            'adm_scenario': page_scenario,
-                            'adm_session_id': adm_session,
-                            'adm_author': survey['results'][page]['admAuthor'],
-                            'adm_alignment_target': survey['results'][page]['admTarget'],
-                            'evalNumber': EVAL_NUMBER
-                        }
+                        document['score'] = res['score']
                         send_document_to_mongo(comparison_collection, document)
                         
                     else:
@@ -110,22 +115,25 @@ def main(mongoDB, EVAL_NUMBER=4):
                                 if x['parameters']['choice'] in probe_ids or x['parameters']['probe_id'] in probe_ids:
                                     probe_responses.append(x['parameters'])
                         found_mini_adm = db_utils.mini_adm_run(EVAL_NUMBER, del_adm_runs_collection, probe_responses, adm_target, survey['results'][page]['admName'])
+                    document = {
+                        'pid': pid,
+                        'adm_type': survey['results'][page]['admAlignment'],
+                        'text_scenario': scenario_id,
+                        'text_session_id': session_id.replace('"', "").strip(),
+                        'adm_scenario': page_scenario,
+                        'adm_session_id': found_mini_adm["session_id"],
+                        'adm_author': survey['results'][page]['admAuthor'],
+                        'adm_alignment_target': survey['results'][page]['admTarget'],
+                        'evalNumber': EVAL_NUMBER
+                    }
+                    if run_new_only and does_doc_exist(comparison_collection, document):
+                        # do not rerun!
+                        continue
                     # get comparison score
                     res = requests.get(f'{ADEPT_URL}api/v1/alignment/compare_sessions?session_id_1={session_id}&session_id_2={found_mini_adm["session_id"]}').json()
                     # send document to mongo
                     if res is not None and 'score' in res:
-                        document = {
-                            'pid': pid,
-                            'adm_type': survey['results'][page]['admAlignment'],
-                            'score': res['score'],
-                            'text_scenario': scenario_id,
-                            'text_session_id': session_id.replace('"', "").strip(),
-                            'adm_scenario': page_scenario,
-                            'adm_session_id': found_mini_adm["session_id"],
-                            'adm_author': survey['results'][page]['admAuthor'],
-                            'adm_alignment_target': survey['results'][page]['admTarget'],
-                            'evalNumber': EVAL_NUMBER
-                        }
+                        document['score'] = res['score']
                         send_document_to_mongo(comparison_collection, document)
                     else:
                         print(f'Error getting comparison for scenarios {scenario_id} and {page_scenario} with text session {session_id} and adm session {found_mini_adm["session_id"]}', res)
@@ -151,6 +159,20 @@ def main(mongoDB, EVAL_NUMBER=4):
                 # get comparison score
                 adm_session_id = tad_most_adm['history'][-1]['parameters']['session_id']
                 res = None
+                document = {
+                    'pid': pid,
+                    'adm_type': 'most aligned',
+                    'text_scenario': scenario_id,
+                    'adm_author': 'TAD',
+                    'attribute': attribute,
+                    'text_session_id': session_id.replace('"', "").strip(),
+                    'adm_session_id': adm_session_id,
+                    'adm_alignment_target': most.get('target', list(most.keys())[0]),
+                    'evalNumber': EVAL_NUMBER
+                }
+                if run_new_only and does_doc_exist(comparison_collection, document):
+                    # do not rerun!
+                    continue
                 if 'Ingroup' in attribute or 'Moral' in attribute:
                     res = requests.get(f'{ADEPT_URL}api/v1/alignment/compare_sessions?session_id_1={session_id}&session_id_2={adm_session_id}').json()
                 else:
@@ -161,18 +183,7 @@ def main(mongoDB, EVAL_NUMBER=4):
                         query_param += f"&session2_probes={probe_id}"
                     res = requests.get(f'{ST_URL}api/v1/alignment/session/subset?{query_param}').json()
                 if res is not None and 'score' in res:
-                    document = {
-                        'pid': pid,
-                        'adm_type': 'most aligned',
-                        'score': res['score'],
-                        'text_scenario': scenario_id,
-                        'adm_author': 'TAD',
-                        'attribute': attribute,
-                        'text_session_id': session_id.replace('"', "").strip(),
-                        'adm_session_id': adm_session_id,
-                        'adm_alignment_target': most.get('target', list(most.keys())[0]),
-                        'evalNumber': EVAL_NUMBER
-                    }
+                    document['score'] = res['score']
                     send_document_to_mongo(comparison_collection, document)
                 else:
                     print(f'Error getting comparison for scenario {scenario_id} with text session {session_id} and adm session {adm_session_id}', res)
@@ -183,6 +194,20 @@ def main(mongoDB, EVAL_NUMBER=4):
             if kit_most_adm is not None:
                 adm_session_id = kit_most_adm['history'][-1]['parameters']['session_id']
                 res = None
+                document = {
+                    'pid': pid,
+                    'adm_type': 'most aligned',
+                    'text_scenario': scenario_id,
+                    'adm_author': 'kitware',
+                    'attribute': attribute,
+                    'text_session_id': session_id.replace('"', "").strip(),
+                    'adm_session_id': adm_session_id,
+                    'adm_alignment_target': most.get('target', list(most.keys())[0]),
+                    'evalNumber': EVAL_NUMBER
+                }
+                if run_new_only and does_doc_exist(comparison_collection, document):
+                    # do not rerun!
+                    continue
                 if 'Ingroup' in attribute or 'Moral' in attribute:
                     res = requests.get(f'{ADEPT_URL}api/v1/alignment/compare_sessions?session_id_1={session_id}&session_id_2={adm_session_id}').json()
                 else:
@@ -193,18 +218,7 @@ def main(mongoDB, EVAL_NUMBER=4):
                         query_param += f"&session2_probes={probe_id}"
                     res = requests.get(f'{ST_URL}api/v1/alignment/session/subset?{query_param}').json()
                 if res is not None and 'score' in res:
-                    document = {
-                        'pid': pid,
-                        'adm_type': 'most aligned',
-                        'score': res['score'],
-                        'text_scenario': scenario_id,
-                        'adm_author': 'kitware',
-                        'attribute': attribute,
-                        'text_session_id': session_id.replace('"', "").strip(),
-                        'adm_session_id': adm_session_id,
-                        'adm_alignment_target': most.get('target', list(most.keys())[0]),
-                        'evalNumber': EVAL_NUMBER
-                    }
+                    document['score'] = res['score']
                     send_document_to_mongo(comparison_collection, document)
                 else:
                     print(f'Error getting comparison for scenario {scenario_id} with text session {session_id} and adm session {adm_session_id}', res)
@@ -219,6 +233,20 @@ def main(mongoDB, EVAL_NUMBER=4):
             if tad_least_adm is not None:
                 adm_session_id = tad_least_adm['history'][-1]['parameters']['session_id']
                 res = None
+                document = {
+                    'pid': pid,
+                    'adm_type': 'least aligned',
+                    'text_scenario': scenario_id,
+                    'adm_author': 'TAD',
+                    'attribute': attribute,
+                    'text_session_id': session_id.replace('"', "").strip(),
+                    'adm_session_id': adm_session_id,
+                    'adm_alignment_target': least.get('target', list(least.keys())[0]),
+                    'evalNumber': EVAL_NUMBER
+                }
+                if run_new_only and does_doc_exist(comparison_collection, document):
+                    # do not rerun!
+                    continue
                 if 'Ingroup' in attribute or 'Moral' in attribute:
                     res = requests.get(f'{ADEPT_URL}api/v1/alignment/compare_sessions?session_id_1={session_id}&session_id_2={adm_session_id}').json()
                 else:
@@ -229,18 +257,7 @@ def main(mongoDB, EVAL_NUMBER=4):
                         query_param += f"&session2_probes={probe_id}"
                     res = requests.get(f'{ST_URL}api/v1/alignment/session/subset?{query_param}').json()
                 if res is not None and 'score' in res:
-                    document = {
-                        'pid': pid,
-                        'adm_type': 'least aligned',
-                        'score': res['score'],
-                        'text_scenario': scenario_id,
-                        'adm_author': 'TAD',
-                        'attribute': attribute,
-                        'text_session_id': session_id.replace('"', "").strip(),
-                        'adm_session_id': adm_session_id,
-                        'adm_alignment_target': least.get('target', list(least.keys())[0]),
-                        'evalNumber': EVAL_NUMBER
-                    }
+                    document['score'] = res['score']
                     send_document_to_mongo(comparison_collection, document)
                 else:
                     print(f'Error getting comparison for scenario {scenario_id} with text session {session_id} and adm session {adm_session_id}', res)
@@ -250,6 +267,20 @@ def main(mongoDB, EVAL_NUMBER=4):
             if kit_least_adm is not None:
                 adm_session_id = kit_least_adm['history'][-1]['parameters']['session_id']
                 res = None
+                document = {
+                    'pid': pid,
+                    'adm_type': 'least aligned',
+                    'text_scenario': scenario_id,
+                    'adm_author': 'kitware',
+                    'attribute': attribute,
+                    'text_session_id': session_id.replace('"', "").strip(),
+                    'adm_session_id': adm_session_id,
+                    'adm_alignment_target': least.get('target', list(least.keys())[0]),
+                    'evalNumber': EVAL_NUMBER
+                }
+                if run_new_only and does_doc_exist(comparison_collection, document):
+                    # do not rerun!
+                    continue
                 if 'Ingroup' in attribute or 'Moral' in attribute:
                     res = requests.get(f'{ADEPT_URL}api/v1/alignment/compare_sessions?session_id_1={session_id}&session_id_2={adm_session_id}').json()
                 else:
@@ -260,18 +291,7 @@ def main(mongoDB, EVAL_NUMBER=4):
                         query_param += f"&session2_probes={probe_id}"
                     res = requests.get(f'{ST_URL}api/v1/alignment/session/subset?{query_param}').json()
                 if res is not None and 'score' in res:
-                    document = {
-                        'pid': pid,
-                        'adm_type': 'least aligned',
-                        'score': res['score'],
-                        'text_scenario': scenario_id,
-                        'adm_author': 'kitware',
-                        'attribute': attribute,
-                        'text_session_id': session_id.replace('"', "").strip(),
-                        'adm_session_id': adm_session_id,
-                        'adm_alignment_target': least.get('target', list(least.keys())[0]),
-                        'evalNumber': EVAL_NUMBER
-                    }
+                    document['score'] = res['score']
                     send_document_to_mongo(comparison_collection, document)
                 else:
                     print(f'Error getting comparison for scenario {scenario_id} with text session {session_id} and adm session {adm_session_id}', res)
@@ -279,6 +299,18 @@ def main(mongoDB, EVAL_NUMBER=4):
 
     print("Human to ADM comparison values added to database.")
 
+
+
+def does_doc_exist(comparison_collection, document):
+    if 'adm_scenario' in document:
+        found_docs = comparison_collection.find({'pid': document['pid'], 'adm_type': document['adm_type'], 'text_scenario': document['text_scenario'], 'adm_scenario': document['adm_scenario'], 'evalNumber': document['evalNumber'],
+                                                'text_session_id': document['text_session_id'], 'adm_session_id': document['adm_session_id'], 'adm_author': document['adm_author'], 'adm_alignment_target': document['adm_alignment_target']})
+    else:
+        found_docs = comparison_collection.find({'pid': document['pid'], 'adm_type': document['adm_type'], 'text_scenario': document['text_scenario'], 'evalNumber': document['evalNumber'],
+                                                'text_session_id': document['text_session_id'], 'adm_session_id': document['adm_session_id'], 'adm_author': document['adm_author'], 'adm_alignment_target': document['adm_alignment_target']})
+    if len(list(found_docs)) > 0:
+        return True
+    return False
 
 
 def send_document_to_mongo(comparison_collection, document):
